UWP: aqua back with xml and audio/multimedia interfaces and new bred library in the back bone: Welcome!! <3tbs, Mummi and bilbo!!
sodapoppin and friends contribution!!

iAssyrian chatting contribution!!

vue_equalizer (fpga_guru) chatting contribution!!

Thiagovgamg chatting contribution!!

Naysayer88 and friends contribution!!

lelandkwong streaming contribution!!

Ash_F0x and JessicaMak streaming contribution!!

WTSRetro/SpiffyDane and Myrkee streaming contribution!!

Ninja and friends streaming contribution!!

erald_guri chatting contribution!!

lasteveq4 C++ interest chat contriubtion!!

cecilphillip and clarkio @"Microsoft Developer" streaming contribution!!

oijtx streaming contribution!!

NorthSeaHero streaming contribution!!

Trainwreckstv and friends streaming contribution!!

-- hi5 contribution...!!

from eddie machine running Windows 10 Pro remotely from bilbo machine running Windows 10 Pro.

+++ aura/node/uwp/directx_application.cpp
@@ -218,8 +218,8 @@ namespace uwp
 
       puiHost->m_pimpl->cast < ::user::interaction_impl >()->m_rectWindowScreen.left = 0;
       puiHost->m_pimpl->cast < ::user::interaction_impl >()->m_rectWindowScreen.top = 0;
-      puiHost->m_pimpl->cast < ::user::interaction_impl >()->m_rectWindowScreen.right = m_window->Bounds.Width;
-      puiHost->m_pimpl->cast < ::user::interaction_impl >()->m_rectWindowScreen.bottom = m_window->Bounds.Height;
+      puiHost->m_pimpl->cast < ::user::interaction_impl >()->m_rectWindowScreen.right = (LONG) m_window->Bounds.Width;
+      puiHost->m_pimpl->cast < ::user::interaction_impl >()->m_rectWindowScreen.bottom = (LONG) m_window->Bounds.Height;
 
    }
